Timing work, based interval on server framerate, started HM work...
The way total time of a session was calculated before was by taking (Tn
- T0). If a user started a session, then stopped it, then restarted it,
all of the calculations reliant on the total time of the session would
be skewed because of the empty block of data residing between T1 and T2.
All I did was calculate the total time inside the main RawGazePacket
loop of calculateTheGoodStuff() by += it with the delta between each
packet. If the delta is > the sample rate's time interval (meaning we
lost a block of time somewhere), just reset it the interval itself and
tack it on like the session never ended.

Before the assumption was made that the sample rate would be 30Hz. TET
can go up to 60Hz, so I now calculate this constant by asking the server
(nicely) what its framerate actually is. This will have consequences,
however, if a session is split into a 30Hz portion and a 60Hz portion.
In order to calculate accurate results, the same sample rate should be
maintained for the entirety of a session.

Made a cursory attempt at introducing the HeatMap to RawGazeData. Will
probably need to make a way to construct a HeatMap by passing it a List
of data. However, testing it with a loop of just 400 random (x,y) heat
points caused significant hang on the UI thread.